LayoutTests/imported/w3c/web-platform-tests/html/webappapis/microtask-queuing/queue-microtask-exceptions.any-expected.txt

CONSOLE MESSAGE: Error: boo

PASS It rethrows exceptions
